requirements.txt
annotated-types==0.7.0
annoying==0.1.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.65.5
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
Unipath==1.1
urllib3==2.3.0
